feat: Add support for Valkey engine
This change allows setting the `engine` variable, to allow using Valkey
as the cache engine.

No breaking changes are expected, as the default value for the `engine`
variable is set to `redis`. Also, the minimum AWS provider version is
not changed as using Redis does not require using a newer provider.